Changed "int" to "long long"

+++ Expression/ExpressionExercise3.cpp
@@ -2,12 +2,12 @@
 using namespace std;
 
 int main() {
-    int a, b;
+    long long a, b;
     cout << "Bilangan a: "; cin >> a;
     cout << "Bilangan b: "; cin >> b;
 
-    int remainder = a % b;
-    int result = a / b;
+    long long remainder = a % b;
+    long long result = a / b;
 
     cout << "Hasil pembagian: " << result << endl;
     cout << "Hasil sisa pembagian: " << remainder << endl;
